Fix all unsafe calls; update to jieba_rs 0.7
Instead of `static mut ...` as Lazy<T>, use an RwLock with `static ...` variables.

This kind of necessitates updating to jieba_rs 0.7, which made it so that TLIDF took the Jieba
instance for methods that need it, instead of at construction time.

Also use OnceLock to remove dependency on once_cell --- OnceLock was stablized in Rust 1.70.0, on
2023-06-01 (2 months after my last commit).